+++ tests/data/ConfigActiveRecord.php
@@ -4,6 +4,11 @@ namespace yii2tech\tests\unit\config\data;
 
 use yii\db\ActiveRecord;
 
+/**
+ * @property string $id
+ * @property string $value
+ * @property string $group
+ */
 class ConfigActiveRecord extends ActiveRecord
 {
     /**